Corrected issues in SDK WASM build caused by recent refactor. Moved compat layer from client diectory to root
