+++ squire_sdk/src/lib.rs
@@ -40,5 +40,7 @@ pub mod response;
 /// Contains all of the components needed for client-server synchronization
 pub mod sync;
 /// Contains the definition of the actor model used by both the client and server
-#[cfg(any(feature = "server", feature = "client"))]
 pub mod actor;
+
+/// A compatability layer to enable use in both native and WASM platforms
+pub mod compat;
